Engine is now a singleton, reason see below
Reason: Although the engine was intended and desingend to instantiated multiple times
by one User Instance it never really working and supported until now so the switch
should be without further issues. Further design choices can know consider using the
Singleton instance of the Engine but should still try to avoid it as much as possible
because of the issues heavy singleton use can introduce.